+++ src/python/index.ts
@@ -1,2 +1,3 @@
 export * as math from './math/index.ts'
+export * as re from './re/index.ts'
 export * as string from './string/index.ts'