+++ src/python/re/index.ts
@@ -1,0 +1,1 @@
+export { sub } from './sub.ts'